renderer: fix mk1 silence detection
The MK1 has a DC offset that pushes silent samples from 0 up to +0x1000000.

This commit also moves all silence processing to the end of the track.
Previously, the renderer would examine silence for the entire track. In
practice this is unlikely to present a noticeable speedup, but it's good
to do anyways.

Fixes #48